BOCM-95 - moves formatters for responses to endpoint responses to own object
